Don't pass unnecessary information to MailChimp.
When using array_flip on the default list of interests from getInterests() we ended up passing mailchimp the full description as well in our json as the arry keys. which is unnecessary, and makes the requests a lot bigger. So just use array_keys() to gethe keys we need. I blame younger, stupider Isaac for this issue.