Set the subsriber info after we check to see if they are already a member of the list.
This allos the method that sets subscriber info to change the info depending on whether or not they are a member.

This code is messy and relies on the welcome_message flag in weird ways, and should be updated, but that's a bigger job than our immediate needs, so ignoring that for now.